dtype error

+++ src/spatial.py
@@ -45,7 +45,7 @@ def spatial_neighbors(gdf, buffer, data_dir, HSI_pool, model, image_size):
             img_crop = torch.tensor(img_crop,device=model.device, dtype=torch.float32).unsqueeze(0)
             with torch.no_grad():
                 score = model.model(img_crop)
-                score = score.astype(np.float32)
+                score = score
             scores.append(score)
         
         if len(scores) == 0:
